more migrating to direct firestore access
