+++ src/components/admin/CalendarEvent.vue
@@ -192,7 +192,7 @@
 </template>
 
 <script>
-import { functions } from "@/firebase/init";
+import { functions, db } from "@/firebase/init";
 export default {
   name: "CalendarEvent",
   data: () => ({
@@ -255,14 +255,25 @@ export default {
         this.updateEvent();
         console.log("update");
       } else {
-        await functions.httpsCallable("createSchedule")({
+        // createSchedule
+        const admin = require("firebase-admin");
+        const arrayUnion = admin.firestore.FieldValue.arrayUnion;
+        var newEvent = {
           name: this.name,
           description: this.description,
           location: this.location,
           type: this.scheduleType,
           start: this.$refs.startTimePicker.formattedDatetime,
           end: this.$refs.endTimePicker.formattedDatetime
-        });
+        };
+        const eventSchedule = db.collection("admin").doc("schedules");
+        await eventSchedule
+          .update({
+            events: arrayUnion(newEvent)
+          })
+          .catch(function(error) {
+            console.error("Error: ", error);
+          });
         this.events.push({
           name: this.name,
           description: this.description,
@@ -323,14 +334,27 @@ export default {
       this.$refs.calendar.next();
     },
     async deleteEvent() {
-      await functions.httpsCallable("deleteSchedule")({
-        name: this.selectedEvent.name,
-        description: this.selectedEvent.description,
-        location: this.selectedEvent.location,
-        type: this.selectedEvent.scheduleType,
-        start: this.selectedEvent.start,
-        end: this.selectedEvent.end
-      });
+      // deleteSchedule
+      const schedsDb = db.collection("admin").doc("schedules");
+      let info = await schedsDb.get();
+      this.events = info
+        .data()
+        .events.filter(
+          schedule =>
+            schedule.name != this.selectedEvent.name &&
+            schedule.description != this.selectedEvent.description &&
+            schedule.location != this.selectedEvent.location &&
+            schedule.type != this.selectedEvent.scheduleType &&
+            schedule.start != this.selectedEvent.start &&
+            schedule.end != this.selectedEvent.end
+        );
+      await schedsDb
+        .update({
+          events: this.events
+        })
+        .catch(function(error) {
+          console.error("Error: ", error);
+        });
       this.events.splice(this.selectedEventIndex, 1);
       this.selectedOpen = false;
     },
