+++ src/components/admin/MentorTable.vue
@@ -162,7 +162,7 @@
 </template>
 
 <script>
-import { functions } from "@/firebase/init";
+import { db } from "@/firebase/init";
 
 export default {
   name: "MentorTable",
@@ -210,8 +210,16 @@ export default {
       this.selected.forEach(entry => {
         UIDList.push(entry.uid);
       });
-      await functions.httpsCallable("massAcceptMentor")({
-        UIDList: UIDList
+      // massAcceptMentor
+      UIDList.forEach(async uid => {
+        const user = db.collection("users").doc(uid);
+        const application = db.collection("mentors").doc(uid);
+        await user.update({
+          applicationStatus: 4
+        });
+        await application.update({
+          status: 4
+        });
       });
     },
     async rejectApplicants() {
@@ -219,20 +227,40 @@ export default {
       this.selected.forEach(entry => {
         UIDList.push(entry.uid);
       });
-      await functions.httpsCallable("massRejectMentor")({
-        UIDList: UIDList
+      // massRejectMentor
+      UIDList.forEach(async uid => {
+        const user = db.collection("users").doc(uid);
+        const application = db.collection("mentors").doc(uid);
+        await user.update({
+          applicationStatus: 2
+        });
+        await application.update({
+          status: 2
+        });
       });
     },
     async save() {
       if (this.editItem.status == 2) {
-        await functions.httpsCallable("rejectMentor")({
-          uid: this.editItem.uid
+        // rejectMentor
+        const user = db.collection("users").doc(this.editItem.uid);
+        const application = db.collection("mentors").doc(this.editItem.uid);
+        await user.update({
+          applicationStatus: 2
+        });
+        await application.update({
+          status: 2
         });
         Object.assign(this.data[this.editIndex], this.editItem);
         this.close();
       } else if (this.editItem.status == 4) {
-        await functions.httpsCallable("acceptMentor")({
-          uid: this.editItem.uid
+        // acceptMentor
+        const user = db.collection("users").doc(this.editItem.uid);
+        const application = db.collection("mentors").doc(this.editItem.uid);
+        await user.update({
+          applicationStatus: 4
+        });
+        await application.update({
+          status: 4
         });
         Object.assign(this.data[this.editIndex], this.editItem);
         this.close();
