+++ src/components/admin/VolunteerTable.vue
@@ -165,7 +165,7 @@
 </template>
 
 <script>
-import { functions } from "@/firebase/init";
+import { db } from "@/firebase/init";
 
 export default {
   name: "VolunteerTable",
@@ -216,8 +216,16 @@ export default {
       this.selected.forEach(entry => {
         UIDList.push(entry.uid);
       });
-      await functions.httpsCallable("massAcceptVolunteer")({
-        UIDList: UIDList
+      // massAcceptVolunteer
+      UIDList.forEach(async uid => {
+        const user = db.collection("users").doc(uid);
+        const application = db.collection("volunteers").doc(uid);
+        await user.update({
+          applicationStatus: 4
+        });
+        await application.update({
+          status: 4
+        });
       });
     },
     async rejectApplicants() {
@@ -225,20 +233,40 @@ export default {
       this.selected.forEach(entry => {
         UIDList.push(entry.uid);
       });
-      await functions.httpsCallable("massRejectVolunteer")({
-        UIDList: UIDList
+      // massRejectVolunteer
+      UIDList.forEach(async uid => {
+        const user = db.collection("users").doc(uid);
+        const application = db.collection("volunteers").doc(uid);
+        await user.update({
+          applicationStatus: 2
+        });
+        await application.update({
+          status: 2
+        });
       });
     },
     async save() {
       if (this.editItem.status == 2) {
-        await functions.httpsCallable("rejectVolunteer")({
-          uid: this.editItem.uid
+        // rejectVolunteer
+        const user = db.collection("users").doc(this.editItem.uid);
+        const application = db.collection("volunteers").doc(this.editItem.uid);
+        await user.update({
+          applicationStatus: 2
+        });
+        await application.update({
+          status: 2
         });
         Object.assign(this.data[this.editIndex], this.editItem);
         this.close();
       } else if (this.editItem.status == 4) {
-        await functions.httpsCallable("acceptVolunteer")({
-          uid: this.editItem.uid
+        // acceptVolunteer
+        const user = db.collection("users").doc(this.editItem.uid);
+        const application = db.collection("volunteers").doc(this.editItem.uid);
+        await user.update({
+          applicationStatus: 4
+        });
+        await application.update({
+          status: 4
         });
         Object.assign(this.data[this.editIndex], this.editItem);
         this.close();
